+++ src/main/java/kitchenpos/menu/controller/MenuGroupRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.menugroup.controller;
+package kitchenpos.menu.controller;
 
-import kitchenpos.menugroup.bo.MenuGroupBo;
-import kitchenpos.menugroup.model.MenuGroup;
+import kitchenpos.menu.bo.MenuGroupBo;
+import kitchenpos.menu.model.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
