+++ src/test/java/kitchenpos/bo/MenuBoTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.bo;
 
 import kitchenpos.menu.dao.MenuDao;
-import kitchenpos.menugroup.dao.MenuGroupDao;
+import kitchenpos.menu.dao.MenuGroupDao;
 import kitchenpos.menu.dao.MenuProductDao;
 import kitchenpos.product.dao.ProductDao;
 import kitchenpos.menu.bo.MenuBo;
